requirements.txt
scaleway-sdk
slumber
argparse
